refactor(cl): convert CourtTestCase to a mixin
