+++ cl/citations/tests.py
@@ -62,7 +62,7 @@ from cl.citations.unmatched_citations_utils import (
 from cl.citations.utils import (
     make_get_citations_kwargs,
 )
-from cl.lib.test_helpers import CourtTestCase, PeopleTestCase, SearchTestCase
+from cl.lib.test_helpers import PeopleTestCase, SearchTestCase
 from cl.search.documents import ParentheticalGroupDocument
 from cl.search.factories import (
     CitationWithParentsFactory,
@@ -90,6 +90,7 @@ from cl.tests.cases import (
     TestCase,
     TransactionTestCase,
 )
+from cl.tests.mixins import CourtMixin
 from cl.users.factories import UserProfileWithParentsFactory
 
 HYPERSCAN_TOKENIZER = HyperscanTokenizer(cache_dir=".hyperscan")
@@ -583,7 +584,7 @@ class CitationTextTest(TestCase):
             )
 
 
-class RECAPDocumentObjectTest(ESIndexTestCase, TestCase):
+class RECAPDocumentObjectTest(TestCase, ESIndexTestCase):
     # pass
     @classmethod
     def setUpTestData(cls):
@@ -657,7 +658,7 @@ class RECAPDocumentObjectTest(ESIndexTestCase, TestCase):
                 self.assertEqual(citation_obj.depth, depth)
 
 
-class CitationObjectTest(ESIndexTestCase, TestCase):
+class CitationObjectTest(TestCase, ESIndexTestCase):
     fixtures: list = []
 
     @classmethod
@@ -1583,12 +1584,12 @@ class CitationObjectTest(ESIndexTestCase, TestCase):
 
 
 class CitationFeedTest(
-    ESIndexTestCase, CourtTestCase, PeopleTestCase, SearchTestCase, TestCase
+    TestCase, ESIndexTestCase, CourtMixin, PeopleTestCase, SearchTestCase
 ):
     @classmethod
     def setUpTestData(cls) -> None:
-        cls.rebuild_index("search.OpinionCluster")
         super().setUpTestData()
+        cls.rebuild_index("search.OpinionCluster")
         call_command(
             "cl_index_parent_and_child_docs",
             search_type=SEARCH_TYPES.OPINION,
@@ -1638,7 +1639,7 @@ class CitationFeedTest(
         self._tree_has_content(r.content, expected_count)
 
 
-class CitationCommandTest(ESIndexTestCase, TestCase):
+class CitationCommandTest(TestCase, ESIndexTestCase):
     """Test a variety of the ways that find_citations can be called."""
 
     fixtures: list = []
@@ -2348,7 +2349,7 @@ class GroupParentheticalsTest(SimpleTestCase):
     return_value="citations_tests",
 )
 class CitationLookUpApiTest(
-    CourtTestCase, PeopleTestCase, SearchTestCase, TestCase
+    TestCase, CourtMixin, PeopleTestCase, SearchTestCase
 ):
     @classmethod
     def setUpTestData(cls) -> None:
